pyproject.toml
[tool.poetry]
name = "self_study_project"
version = "0.1.0"
description = ""
authors = ["KirillOreshchuk <oreshchuk.kirill@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
flake8 = "^7.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
